Enable user to format the current file
Now the user of the plugin can choose the formatted version of the
current file content in the picker. This then replaces the content of
the current buffer by the formatted file contents. The buffer is not
saved however, hence the user can still choose to revert the actions by
choosing the 'orgiginal' picker entry until they save the file.