Implement interactive diff patching
Upon selection of the 'diff' entry in the picker, users are now prompted
to edit the diff they want to apply to the current buffer. Using `:W` in
normal mode in that buffer the current state of the diff is applied to
the underlying buffer containing the file that is to be formatted. Using
the regular `:q` users can enter this new prompt buffer. This action
also aborts any changes done after the last `:W`.